third_party/WebKit/LayoutTests/fast/canvas/canvas-overloads-drawImageFromRect-expected.txt

Test the behavior of CanvasRenderingContext2D.drawImageFromRect() when called with different numbers of arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.drawImageFromRect() threw exception TypeError: Failed to execute 'drawImageFromRect' on 'CanvasRenderingContext2D': 1 argument required, but only 0 present..
PASS ctx.drawImageFromRect(imageElement) is undefined
PASS ctx.drawImageFromRect(imageElement, 0) is undefined
PASS ctx.drawImageFromRect(imageElement, 0, 0) is undefined
PASS ctx.drawImageFromRect(imageElement, 0, 0, 0) is undefined
PASS ctx.drawImageFromRect(imageElement, 0, 0, 0, 0) is undefined
PASS ctx.drawImageFromRect(imageElement, 0, 0, 0, 0, 0) is undefined
PASS ctx.drawImageFromRect(imageElement, 0, 0, 0, 0, 0, 0) is undefined
PASS ctx.drawImageFromRect(imageElement, 0, 0, 0, 0, 0, 0, 0) is undefined
PASS ctx.drawImageFromRect(imageElement, 0, 0, 0, 0, 0, 0, 0, 0) is undefined
PASS ctx.drawImageFromRect(imageElement, 0, 0, 0, 0, 0, 0, 0, 0, 0) is undefined
PASS ctx.drawImageFromRect(imageElement, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0) is undefined
PASS ctx.drawImageFromRect(imageElement, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0) is undefined
PASS ctx.drawImageFromRect(imageElement, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0) is undefined
PASS successfullyParsed is true

TEST COMPLETE
